Consolidate view_file and code_theme_toggle into view_file folder
