+++ src/components/project_view/submission_detail/ag_test_suite_setup_result_detail.vue
@@ -75,7 +75,7 @@ import {
     Submission
 } from "ag-client-typescript";
 
-import ViewFile from "@/components/view_file.vue";
+import ViewFile from "@/components/view_file/view_file.vue";
 import { handle_global_errors_async } from '@/error_handling';
 
 @Component({
