+++ src/demos/ui_demos.vue
@@ -100,7 +100,7 @@ import { Component, Vue } from 'vue-property-decorator';
 
 import Dropdown from '@/components/dropdown.vue';
 import Toggle from '@/components/toggle.vue';
-import ViewFile from '@/components/view_file.vue';
+import ViewFile from '@/components/view_file/view_file.vue';
 
 import ButtonDemo from './button_demo.vue';
 import ColorPaletteDemo from './color_palette_demo.vue';
